test(workout): add workout service integration tests
- 16 integration tests covering full service → JPA → PostgreSQL stack
- Verifies: start, getById, listByUser, addExercise (with sets), complete, delete
- Confirms persistence round-trip: DB-assigned IDs, completedAt, nested sets
- Confirms ownership enforcement and lifecycle guards against real DB
- Fix V13 migration: rpe column changed from NUMERIC(3,1) to INTEGER
  (Hibernate schema validation expected integer type for Integer field)
- Seeds a test exercise per test (FK constraint) and cleans up in @AfterEach

+++ src/main/resources/db/changelog/V13__create_workout_sets_table.sql
@@ -8,7 +8,7 @@ CREATE TABLE workout_sets (
     reps                INTEGER        NOT NULL,
     weight_value        NUMERIC(10, 4) NOT NULL,
     weight_unit         VARCHAR(10)    NOT NULL,
-    rpe                 NUMERIC(3, 1),
+    rpe                 INTEGER,
     CONSTRAINT pk_workout_sets               PRIMARY KEY (id),
     CONSTRAINT fk_workout_sets_we            FOREIGN KEY (workout_exercise_id) REFERENCES workout_exercises (id) ON DELETE CASCADE
 );
